update field types

+++ stats/stats-dto/src/main/java/ru/practicum/dto/ViewStatsDtoMapper.java
@@ -10,7 +10,7 @@ public class ViewStatsDtoMapper {
         return new ViewStats();
     }
 
-    public static ViewStatsDto toViewStatsDto(ViewStats viewStats, int hits) {
+    public static ViewStatsDto toViewStatsDto(ViewStats viewStats, Long hits) {
         ViewStatsDto viewStatsDto = new ViewStatsDto();
         viewStatsDto.setHits(hits);
         viewStatsDto.setApp(viewStats.getApp());
